dhtml sample integration
Added the files for HTML, CSS, and JS within their associated folder
(it's under dhtmlxSample). Integrated the html code for the calendar in
viewCalendarMainAdmin.html. It's rendering the pop up like in the
sample, but the calendar doesn't appear (maybe the clash of html code
with Dom and mine). Currently working on this.